patcher/pyproject.toml
[tool.poetry]
name = "patcher"
version = "0.1.0"
description = ""
authors = ["Jeffrey04 <mycyberpet@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.1"
ujson = "^4.0.1"
PyMySQL = "^0.10.1"
crawler = "^0.0.2"
cryptography = "^3.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
patcher = "patcher.patcher:main"